requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
beautifulsoup4
arxiv
pymupdf
wikipedia

langchain-text-splitters
langchain-openai
chromadb

langchain-community

sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma

langchain-openai

streamlit

langchain_groq

fastapi
uvicorn
langserve
sse_starlette